+++ src/lib/childWindowCommandQueue.ts
@@ -1,0 +1,75 @@
+import type { ChildWindowCommandName } from "./childWindowProtocol";
+
+export interface ChildWindowCommandEnvelope {
+  event: ChildWindowCommandName;
+  payload: unknown;
+}
+
+interface ChildWindowCommandState {
+  token: string;
+  expectedEvent: ChildWindowCommandName;
+  ready: boolean;
+  pending: ChildWindowCommandEnvelope[];
+}
+
+/**
+ * 父窗口只在内存中保存尚未被子页面消费的命令。状态以窗口 label 和 ready token
+ * 共同隔离，避免已销毁 WebView 的迟到事件释放新窗口队列。
+ */
+export class ChildWindowCommandQueue {
+  private readonly states = new Map<string, ChildWindowCommandState>();
+
+  register(label: string, token: string, expectedEvent: ChildWindowCommandName) {
+    const current = this.states.get(label);
+    if (current?.token === token && current.expectedEvent === expectedEvent) return;
+
+    this.states.set(label, {
+      token,
+      expectedEvent,
+      ready: false,
+      pending: [],
+    });
+  }
+
+  dispatch(
+    label: string,
+    event: ChildWindowCommandName,
+    payload: unknown,
+  ): ChildWindowCommandEnvelope[] {
+    const command = { event, payload };
+    const state = this.states.get(label);
+    if (!state || state.expectedEvent !== event || state.ready) {
+      return [command];
+    }
+
+    state.pending.push(command);
+    return [];
+  }
+
+  markReady(
+    label: string,
+    token: string,
+    event: ChildWindowCommandName,
+  ): ChildWindowCommandEnvelope[] {
+    const state = this.states.get(label);
+    if (!state || state.token !== token || state.expectedEvent !== event) return [];
+    state.ready = true;
+    return state.pending.splice(0);
+  }
+
+  markLoading(label: string, token: string) {
+    const state = this.states.get(label);
+    if (!state || state.token !== token) return false;
+
+    state.ready = false;
+    return true;
+  }
+
+  markFailed(label: string, token: string) {
+    return this.markLoading(label, token);
+  }
+
+  clear(label: string) {
+    this.states.delete(label);
+  }
+}
