+++ src/main.tsx
@@ -7,9 +7,6 @@ import "@fontsource/inter/400.css";
 import "@fontsource/inter/500.css";
 import "@fontsource/inter/600.css";
 import "@fontsource-variable/noto-sans-sc";
-import "./i18n";
-import ErrorBoundary from "./components/ErrorBoundary";
-import { Toaster } from "./components/ui/sonner";
 import "./index.css";
 import {
   applyThemeToDOM,
@@ -17,9 +14,13 @@ import {
   THEME_SNAPSHOT_CACHE_KEY,
   ThemeProvider,
 } from "./context/ThemeContext";
+import {
+  scheduleChildWindowShellReady,
+  signalChildWindowLoadFailed,
+  signalChildWindowLoadStarted,
+} from "./lib/childWindowLifecycle";
 import { DEFAULT_THEME_ID, themes } from "./lib/themes";
 import { installWebviewReloadGuard } from "./lib/webviewReloadGuard";
-import { scheduleChildWindowReady } from "./lib/windowManager";
 
 // Apply cached theme synchronously before React renders to avoid flash
 try {
@@ -45,6 +46,7 @@ const params = new URLSearchParams(window.location.search);
 const windowType = params.get("window");
 
 if (windowType) {
+  void signalChildWindowLoadStarted().catch(() => {});
   // Child window: lightweight provider stack, no full App
   // These entry points are independent and should load in parallel; serial awaits would add an
   // unnecessary chunk round trip to every child-window open.
@@ -60,31 +62,78 @@ if (windowType) {
       <span className="size-5 animate-spin rounded-full border-2 border-primary border-t-transparent" />
     </div>,
   );
-  scheduleChildWindowReady();
+  scheduleChildWindowShellReady();
 
-  const [{ ChildAppProvider }, { default: ChildWindowRouter }] = await Promise.all([
-    import("./context/ChildAppProvider"),
-    import("./ChildWindowRouter"),
-  ]);
+  try {
+    const [
+      { ChildAppProvider },
+      { default: ChildWindowRouter },
+      { default: ErrorBoundary },
+      { Toaster },
+    ] = await Promise.all([
+      import("./context/ChildAppProvider"),
+      import("./ChildWindowRouter"),
+      import("./components/ErrorBoundary"),
+      import("./components/ui/sonner"),
+    ]);
 
-  childRoot.render(
-    <React.StrictMode>
-      <ErrorBoundary>
-        <ChildAppProvider>
-          <ThemeProvider>
-            <ChildWindowRouter windowType={windowType} />
-            <Toaster />
-          </ThemeProvider>
-        </ChildAppProvider>
-      </ErrorBoundary>
-    </React.StrictMode>,
-  );
+    childRoot.render(
+      <React.StrictMode>
+        <ErrorBoundary>
+          <ChildAppProvider>
+            <ThemeProvider>
+              <ChildWindowRouter windowType={windowType} />
+              <Toaster />
+            </ThemeProvider>
+          </ChildAppProvider>
+        </ErrorBoundary>
+      </React.StrictMode>,
+    );
+  } catch {
+    void signalChildWindowLoadFailed("bootstrap-import").catch(() => {});
+    let errorTitle = "Something went wrong";
+    let reloadLabel = "Reload";
+    try {
+      const { default: i18n } = await import("./i18n");
+      errorTitle = i18n.t("error.somethingWentWrong");
+      reloadLabel = i18n.t("error.reloadApplication");
+    } catch {}
+    childRoot.render(
+      <div
+        className="flex h-screen w-full items-center justify-center bg-background p-8"
+        style={{ backgroundColor: "var(--df-bg, #0d1117)" }}
+        role="alert"
+      >
+        <div className="max-w-md text-center">
+          <p className="text-destructive">{errorTitle}</p>
+          <button
+            type="button"
+            className="mt-6 rounded-md bg-primary px-4 py-2 text-primary-foreground"
+            onClick={() => window.location.reload()}
+          >
+            {reloadLabel}
+          </button>
+        </div>
+      </div>,
+    );
+  }
 } else {
   // Main window: full app with all providers
-  const { getCurrentWindow } = await import("@tauri-apps/api/window");
-  const { setOwnerMainWindowLabel } = await import("./lib/windowManager");
-  const { AppProvider } = await import("./context/AppContext");
-  const { default: App } = await import("./App");
+  const [
+    { getCurrentWindow },
+    { setOwnerMainWindowLabel },
+    { AppProvider },
+    { default: App },
+    { default: ErrorBoundary },
+    { Toaster },
+  ] = await Promise.all([
+    import("@tauri-apps/api/window"),
+    import("./lib/windowManager"),
+    import("./context/AppProvider"),
+    import("./App"),
+    import("./components/ErrorBoundary"),
+    import("./components/ui/sonner"),
+  ]);
   setOwnerMainWindowLabel(getCurrentWindow().label);
 
   ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
