+++ src/pages/RemoteFileEditorPage.tsx
@@ -1,7 +1,6 @@
 import { closeSearchPanel } from "@codemirror/search";
 import { EditorState } from "@codemirror/state";
 import { EditorView } from "@codemirror/view";
-import { listen } from "@tauri-apps/api/event";
 import { join, tempDir } from "@tauri-apps/api/path";
 import { getCurrentWindow } from "@tauri-apps/api/window";
 import { openPath } from "@tauri-apps/plugin-opener";
@@ -36,6 +35,8 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
 import { useApp } from "@/context/AppContext";
+import { useChildWindowCommand } from "@/hooks/useChildWindowCommand";
+import { CHILD_WINDOW_COMMANDS } from "@/lib/childWindowProtocol";
 import {
   type CursorPosition,
   codeMirrorFileViewExtensions,
@@ -317,23 +318,14 @@ export default function RemoteFileEditorPage() {
     void loadFile(tabId(initialData));
   }, [initialData, loadFile]);
 
-  useEffect(() => {
-    const currentWindow = getCurrentWindow();
-    let unlisten: (() => void) | undefined;
-
-    listen<RemoteFileEditorOpenPayload>("remote-file-editor-open", (event) => {
-      if (event.payload.targetLabel && event.payload.targetLabel !== currentWindow.label) return;
-      addOrFocusTab(event.payload.data);
-    })
-      .then((dispose) => {
-        unlisten = dispose;
-      })
-      .catch(() => {});
-
-    return () => {
-      unlisten?.();
-    };
-  }, [addOrFocusTab]);
+  useChildWindowCommand<RemoteFileEditorOpenPayload>(
+    CHILD_WINDOW_COMMANDS.remoteFileEditorOpen,
+    (payload) => {
+      const currentWindow = getCurrentWindow();
+      if (payload.targetLabel && payload.targetLabel !== currentWindow.label) return;
+      addOrFocusTab(payload.data);
+    },
+  );
 
   useEffect(() => {
     const currentWindow = getCurrentWindow();
